requirements.txt
# Core
feedparser>=6.0
httpx>=0.27
pyyaml>=6.0
python-dotenv>=1.0

# Database
psycopg2-binary>=2.9

# LLM
google-genai>=1.0

# Scheduler
apscheduler>=3.10,<4.0

# Testing
pytest>=8.0
